# kfone_admin_backend/swagger.yaml
openapi: 3.0.0
info:
  title: Device Service API
  description: API for managing electronic devices and promotions
  version: 1.0.0

paths:
  /devices:
    get:
      summary: Retrieve a list of all devices
      description: <b>Scope required:</b> `devices_list`
      tags:
        - devices
      responses:
        '200':
          description: A list of all devices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Device'
    post:
      summary: Add a new device
      description: <b>Scope required:</b> `devices_add`
      tags:
        - devices
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceInput'
      responses:
        '201':
          description: The newly created device
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'

  /devices/search/{name}:
    get:
      summary: Search a device by name
      description: <b>Scope required:</b> `devices_list`
      tags:
        - devices
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
          description: Name of the device to search
      responses:
        '200':
          description: The requested device
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '404':
          description: Device not found

  /devices/{device_id}:
    parameters:
      - in: path
        name: device_id
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
        description: ID of the device to retrieve or modify
    get:
      summary: Retrieve a device by ID
      description: <b>Scope required:</b> `devices_list`
      tags:
        - devices
      responses:
        '200':
          description: The requested device
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '404':
          description: Device not found
    put:
      summary: Update a device
      description: <b>Scope required:</b> `devices_modify`
      tags:
        - devices
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceInput'
      responses:
        '200':
          description: The updated device
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '404':
          description: Device not found
    delete:
      summary: Remove a device
      description: <b>Scope required:</b> `devices_delete`
      tags:
        - devices
      responses:
        '204':
          description: Device removed
        '404':
          description: Device not found

  /promotions:
    get:
      summary: Get all promotions
      description: <b>Scope required:</b> `promotions_list`
      operationId: getPromotions
      tags:
        - promotions
      responses:
        '200':
          description: List of promotions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Promotion'
        '404':
          description: No promotions found

    post:
      summary: Add new promotion
      description: <b>Scope required:</b> `promotions_add`
      operationId: addPromotion
      tags:
        - promotions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromotionInput'
      responses:
        '201':
          description: The newly created promotion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Promotion'
        '400':
          description: Invalid input
        '409':
          description: Promotion with provided promo_code already exists

  /promotions/{promo_id}:
    parameters:
      - in: path
        name: promo_id
        required: true
        schema:
          type: integer
        description: Numeric ID of the promotion to get or update
    get:
      summary: Get promotion by ID
      description: <b>Scope required:</b> `promotions_list`
      operationId: getPromotionById
      tags:
        - promotions
      responses:
        '200':
          description: The requested promotion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Promotion'
        '404':
          description: Promotion not found

    put:
      summary: Update promotion by ID
      description: <b>Scope required:</b> `promotions_modify`
      operationId: updatePromotionById
      tags:
        - promotions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromotionInput'
      responses:
        '200':
          description: The updated promotion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Promotion'
        '400':
          description: Invalid input
        '404':
          description: Promotion not found

    delete:
      summary: Delete promotion by ID
      description: <b>Scope required:</b> `devices_list`
      operationId: deletePromotionById
      tags:
        - promotions
      responses:
        '204':
          description: Promotion deleted
        '404':
          description: Promotion not found

  /promotions/devices:
    post:
      summary: Add promotion to devices
      description: <b>Scope required:</b> `promotions_modify`
      operationId: addPromotionToDevice
      tags:
        - promotions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddPromotionToDeviceInput'
      responses:
        '200':
          description: The updated list of devices with added promotion
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Device'
        '400':
          description: Invalid input
        '404':
          description: Promotion or device not found
          
  /customers:
    get:
      summary: Get all customers
      description: <b>Scope required:</b> `customers_list`
      tags:
        - customers
      responses:
        '200':
          description: Returns a list of all customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Customer'
    post:
      summary: Add a new custome
      description: <b>Scope required:</b> `customers_add`
      tags:
        - customers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerCreateInput'
      responses:
        '201':
          description: Returns the newly created customer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
  /customers/{customerId}:
    parameters:
      - name: customerId
        in: path
        description: ID of the customer to retrieve or update
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get a customer by ID
      description: <b>Scope required:</b> `customers_list`
      tags:
        - customers
      responses:
        '200':
          description: Returns the customer with the given ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '404':
          description: Customer not found
    patch:
      summary: Update an existing customer
      description: <b>Scope required:</b> `customers_modify`
      tags:
        - customers
      parameters:
        - name: customerId
          in: path
          description: ID of the customer to update
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerPatchInput'
      responses:
        '200':
          description: Returns the updated customer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '404':
          description: Customer not found
    delete:
      summary: Delete a customer by ID
      description: <b>Scope required:</b> `customers_delete`
      tags:
        - customers
      parameters:
        - name: customerId
          in: path
          description: ID of the customer to delete
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Customer deleted successfully
        '404':
          description: Customer not found


  /sales:
    get:
      summary: "Sales Trends API"
      description: <b>Scope required:</b> `sales_trends_view`
      operationId: getSalesTrend
      tags:
        - sales
      responses:
        '200':
          description: List of sales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Promotion'
        '404':
          description: No promotions found

components:
  schemas:
    Device:
      type: object
      properties:
        device_id:
          type: integer
          format: int64
        name:
          type: string
        image_uri:
          type: string
        qty:
          type: integer
          format: int32
        description:
          type: string
        price:
          type: number
          format: float
        promo_id:
          type: array
          items:
            type: integer
            format: int64
      required:
        - device_id
        - name
        - qty
        - description
        - price

    DeviceInput:
      type: object
      properties:
        name:
          type: string
        image_uri:
          type: string
        qty:
          type: integer
          format: int32
        description:
          type: string
        price:
          type: number
          format: float

    Promotion:
      type: object
      properties:
        promo_id:
          type: integer
          format: int64
        promo_code:
          type: string
        discount:
          type: number
          format: float
        tier:
          type: array
          items:
            type: string
            enum: [Silver, Bronze, Platinum, None]
        devices:
          type: array
          items:
            type: integer
            format: int64
      required:
        - promo_id
        - promo_code
        - discount


    PromotionInput:
      type: object
      properties:
        promo_code:
          type: string
        discount:
          type: number
          format: float
        tier:
          type: array
          items:
            type: string
            enum: [Silver, Bronze, Platinum, None]

    AddPromotionToDeviceInput:
      type: object
      required:
        - promo_code
        - device_ids
      properties:
        promo_code:
          type: string
          description: The code of the promotion to add
        device_ids:
          type: array
          items:
            type: integer
          description: The IDs of the devices to add the promotion to
    
    Customer:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the customer
        firstName:
          type: string
          description: First name of the customer
        lastName:
          type: string
          description: Last name of the customer
        email:
          type: string
          description: Email Address of the customer
        tier:
          type: string
          enum: [Bronze, Silver, Platinum, None]
          description: Customer's tier level
        country:
          type: string
          description: Country where the customer is located
      required:
        - id
        - firstName
        - lastName
        - email
        - tier
        - country
    CustomerCreateInput:
      type: object
      properties:
        firstName:
          type: string
          description: First name of the customer
        lastName:
          type: string
        tier:
          type: string
          enum: [Bronze, Silver, Platinum, None]
          default: None
          description: Customer's tier level
        country:
          type: string
          description: Country where the customer is located
        email:
          type: string
          description: Email Address of the customer
      required:
        - firstName
        - lastName
        - tier
        - country
    CustomerPatchInput:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        tier:
          type: string
          enum: [Bronze, Silver, Platinum, None]
          description: Customer's tier level